Terminal

+++ prog-maven/tema12-ficheros/src/main/java/content/Terminal/MiniFileManager.java
@@ -179,6 +179,7 @@ class MiniFileManager {
      */
     public void remove(String rutaString) throws FileManagerException {
         File ruta = new File(MiniTerminal.getDirectorioActual(), rutaString);
+        boolean tieneSubcarpetas = false;
         if (!ruta.exists())
             throw new FileManagerException("No existe el archivo o direcorio");
 
@@ -186,13 +187,14 @@ class MiniFileManager {
             ruta.delete();
             return;
         }
+        if (ruta.isDirectory()) {
 
-        boolean tieneSubcarpetas = false;
-        for (File f : ruta.listFiles()) {
-            if (f.isDirectory())
-                tieneSubcarpetas = true;
-            else
-                f.delete();
+            for (File f : ruta.listFiles()) {
+                if (f.isDirectory())
+                    tieneSubcarpetas = true;
+                else
+                    f.delete();
+            }
         }
 
         if (tieneSubcarpetas)
@@ -253,8 +255,8 @@ class MiniFileManager {
         File file = new File(MiniTerminal.getDirectorioActual(), fichero);
         Path ruta = Paths.get(file.toURI());
 
-        if (!Files.exists(ruta))
-            throw new FileManagerException("No es un fichero");
+        if (!Files.exists(ruta) || Files.isDirectory(ruta))
+            throw new FileManagerException("No es un fichero válido");
 
         try {
 
@@ -278,8 +280,8 @@ class MiniFileManager {
         File file = new File(MiniTerminal.getDirectorioActual(), fichero);
         Path ruta = Paths.get(file.toURI());
 
-        if (!Files.exists(ruta))
-            throw new FileManagerException("No es un fichero");
+        if (!Files.exists(ruta) || Files.isDirectory(ruta))
+            throw new FileManagerException("No es un fichero válido");
 
         try {
